+++ ui/src/components/header/Header.tsx
@@ -5,10 +5,12 @@ import logoPath from '../../assets/photoview-logo.svg'
 import { authToken } from '../../helpers/authentication'
 
 const Header = () => (
-  <div className="bg-white flex items-center py-3 px-4 shadow-separator">
-    <h1 className="mr-4 flex-shrink-0">
-      <img className="h-12" src={logoPath} alt="logo" />
-      <span className="hidden">Photoview</span>
+  <div className="bg-white flex items-center justify-between py-3 px-4 lg:px-8 lg:pt-4 shadow-separator lg:shadow-none">
+    <h1 className="mr-4 lg:mr-8 flex-shrink-0 flex items-center">
+      <img className="h-12 lg:h-10" src={logoPath} alt="logo" />
+      <span className="hidden lg:block ml-2 text-2xl font-light">
+        Photoview
+      </span>
     </h1>
     {authToken() ? <SearchBar /> : null}
   </div>
